Updates widgetHandling to ChartHandling composable
