+++ aas-web-ui/src/components/Widgets/Histogram.vue
@@ -39,13 +39,13 @@
     import { defineComponent } from 'vue';
     import { useRoute } from 'vue-router';
     import { useTheme } from 'vuetify';
+    import { useChartHandling } from '@/composables/ChartHandling';
     import DashboardHandling from '@/mixins/DashboardHandling';
-    import WidgetHandling from '@/mixins/WidgetHandling';
     import { useNavigationStore } from '@/store/NavigationStore';
 
     export default defineComponent({
         name: 'Histogram',
-        mixins: [WidgetHandling, DashboardHandling],
+        mixins: [DashboardHandling],
         props: ['chartData', 'timeVariable', 'yVariables', 'chartOptionsExternal', 'editDialog'],
 
         setup() {
@@ -53,10 +53,13 @@
             const navigationStore = useNavigationStore();
             const route = useRoute();
 
+            const { prepareHistogramData } = useChartHandling();
+
             return {
                 theme, // Theme Object
                 navigationStore,
                 route, // Route Object
+                prepareHistogramData,
             };
         },
